requirements.txt
tk
Pillow
pytesseract
groq
keyboard
win32gui
win32ui
comtypes
numpy
python-dotenv
